Move the charge transfer tests into test_chargetransfer.py
